requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas
ca-certificates
certifi
cffi
colorama
freetype
intel-openmp
jpeg
lcms2
lerc
#libcxx
libdeflate
libffi
libpng
libtiff
libwebp-base
lz4-c=1.9.4
mkl=2019.4
mkl-service
mkl_fft
mkl_random
#ncurses
ninja
ninja-base
numpy
numpy-base
olefile
mmpose
#opencv-python
openjpeg
openssl
pandas
pillow
pip
pycparser
python
python-dateutil
pytorch
pytz
#readline
setuptools
six
sqlite
tk
torchvision
tqdm
wheel
xz
zlib
zstd
